adding a map to the user settings
there is a map in the user settings that should be able to keep track of the budget for all the categories that will exist, there is a bug though that the map doesn't get saved into the database. another issue at this point is that the base categories are not accessible to have custom categories.